backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
python-multipart==0.0.9
requests==2.32.3
pypdf==4.3.1
python-docx==1.1.2
